# docs/P3/openapi.yaml
openapi: 3.0.3
info:
  title: CampusHub API
  version: 1.0.0
  description: API for CampusHub campus mutual-aid platform (subset for P3 deliverable)
servers:
  - url: /api/v1
paths:
  /auth/register:
    post:
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: 请求参数错误（校验失败）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 冲突（例如学号已注册）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/email-code:
    post:
      summary: 发送邮箱验证码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailCodeRequest'
      responses:
        '200':
          description: 验证码已发送
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 冲突（邮箱已绑定或发送过于频繁）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: 用户登录，返回 JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 认证失败（学号或密码错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /orders/{orderId}:
    get:
      summary: 查看订单详情
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未找到
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{orderId}/reviews:
    post:
      summary: 提交评价
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '200':
          description: 评价提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 权限不足（非订单参与方或订单未完成）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/demands/{demandId}/review:
    post:
      summary: 管理员审核需求（通过/拒绝）
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: demandId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [approve,reject]
                reason:
                  type: string
      responses:
        '200':
          description: 审核完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandResponse'
        '403':
          description: 权限不足（需管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users:
    get:
      summary: 管理员查询用户列表
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
  /admin/users/{userId}/ban:
    post:
      summary: 管理员封禁用户
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: 封禁成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users/{userId}/unban:
    post:
      summary: 管理员解封用户
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 解封成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        使用 JWT Bearer Token。Token 应包含用户标识与权限信息（例如 `roles` 或 `role` 声明），管理员接口需包含 `admin` 角色。
  schemas:
    StandardResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
      properties:
        errors:
          type: object
    UserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        studentId:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          description: 用户角色
        status:
          type: string
          description: 用户状态
        creditScore:
          type: integer
          description: 信用分
        balance:
          type: number
          format: double
          description: 可用余额
        frozenBalance:
          type: number
          format: double
          description: 冻结金额
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
      properties:
        data:
          $ref: '#/components/schemas/UserSummary'
    RegisterRequest:
      type: object
      required: [email,verificationCode,studentId,password]
      properties:
        email:
          type: string
          format: email
          description: 学校邮箱
        verificationCode:
          type: string
          minLength: 4
          maxLength: 8
        studentId:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 8
          description: 密码最少 8 字符，建议使用复杂密码
        nickname:
          type: string
          maxLength: 64
        avatarUrl:
          type: string
          format: uri
    EmailCodeRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        studentId:
          type: string
          minLength: 3
          maxLength: 64
    LoginRequest:
      type: object
      required: [loginId,password]
      properties:
        loginId:
          type: string
          description: 邮箱或学号
        password:
          type: string
          minLength: 8
    AuthData:
      type: object
      properties:
        token:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/UserSummary'
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
      properties:
        data:
          $ref: '#/components/schemas/AuthData'
...
    Demand:
      type: object
      properties:
        id:
          type: string
        publisherId:
          type: string
        publisherDisplayName:
          type: string
        publisher:
          $ref: '#/components/schemas/UserSummary'
        title:
          type: string
        description:
          type: string
        category:
          type: string
        campusZone:
          type: string
        location:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        reward:
          type: number
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        anonymous:
          type: boolean
        anonymousCode:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DemandResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
      properties:
        data:
          $ref: '#/components/schemas/Demand'
...
    Order:
      type: object
      properties:
        orderId:
          type: string
        demand:
          $ref: '#/components/schemas/Demand'
        provider:
          $ref: '#/components/schemas/UserSummary'
        requester:
          $ref: '#/components/schemas/UserSummary'
        status:
          type: string
        createdAt:
          type: string
    OrderResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
      properties:
        data:
          $ref: '#/components/schemas/Order'
    ReviewCreate:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 1000
    Review:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        authorId:
          type: string
        targetId:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
    ReviewResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
      properties:
        data:
          $ref: '#/components/schemas/Review'
...
    UserListResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/UserSummary'
            page:
              type: integer
            size:
              type: integer
            total:
              type: integer
...